* [jsfm] move setup to a seperate file
Default weex project will setup frameworks under `../../frameworks/index`. In some custom case, you can use whatever you want by pass the frameworks as arguments to  this `setup` function.
